Port the build system to Bazel
`bpf-linker` has a fairly complex build process that depends on a
matching Rust compiler, LLVM build, target C and C++ runtimes, BPF
linker, and BTF inspection tool. The Cargo build and the existing
release jobs assemble those dependencies separately on each runner,
which makes cross-compilation and release reproduction depend on
runner state.

This PR adds Bazel build configuration to support fully hermetic
builds that can be remotely executed and cross-compiled. Along the way
we discovered that some build tools such as btfdump, the rust compiler,
and the musl CC toolchain were not fully pinned. These defects are
corrected in the Bazel build.

In the non-bazel build, `compiletest.rs` is responsible for compiling
the BTF/BPF fixtures used in integration testing. In the bazel build,
these compilations are lifted into the first-class build graph and
expressed in BUILD.bazel. This leads to duplication of the per-fixture
build flags between each fixture's file header and the BUILD.bazel; an
AGENTS.md/prompt are provided for a bot to keep them in sync.

The release support matrix is expanded to cover Windows x86/arm in
addition to the existing Linux musl and Macos targets; future PRs will
further tweak the release process to upload split debuginfo
(dsym/pdb/elf) to reduce download sizes.

We now support remote execution; Buildbuddy is recommended. CI is
automatically configured to dispatch actions remotely when running
`bazel test --config=remote`, contributors can do the same. It is
recommended to provide a buildbuddy API key, see
https://app.buildbuddy.io/docs/setup/ for instructions.

A few followup items are left to future PRs:
- Simplify the rules_rs source-linker patch after bpf-linker publishes
  Windows binaries
- Merge the btfdump Cargo workspace into the root Cargo.toml dependency
  closure
- use core-only no_std after rules_rust supports it
- remove the archiver and allocator workarounds after their upstream
  fixes
- package Windows PDB and Linux split debug information
- consider reenabling Windows CI if it can be optimized to a reasonable
  runtime
