+++ .bazelversion
@@ -1,0 +1,1 @@
+9.1.1
