+++ .gitignore
@@ -1,2 +1,5 @@
 /target
+/bazel-*
 .vscode/settings.json
+*.swp
+.bazelrc.user
