+++ bazel/BUILD.bazel
@@ -1,0 +1,4 @@
+exports_files([
+    "defs.bzl",
+    "release.bzl",
+])
