+++ bazel/btfdump/BUILD.bazel
@@ -1,0 +1,4 @@
+exports_files([
+    "Cargo.lock",
+    "Cargo.toml",
+])
