+++ bazel/btfdump/src/lib.rs
@@ -1,0 +1,1 @@
+// This package exists only to resolve the Bazel test tool dependency.
