+++ src/linker.rs
@@ -1,3 +1,5 @@
+#[cfg(unix)]
+use std::os::unix::ffi::OsStrExt as _;
 use std::{
     borrow::Cow,
     collections::HashSet,
@@ -5,7 +7,6 @@ use std::{
     fs,
     io::{self, Read as _},
     ops::Deref,
-    os::unix::ffi::OsStrExt as _,
     path::{Path, PathBuf},
     str::{self, FromStr},
 };
@@ -16,6 +17,16 @@ use thiserror::Error;
 use tracing::{debug, error, info, warn};
 
 use crate::llvm::{self, LLVMContext, LLVMModule, LLVMTargetMachine, MemoryBuffer};
+
+#[cfg(unix)]
+fn os_str_from_bytes(bytes: &[u8]) -> Result<&OsStr, str::Utf8Error> {
+    Ok(OsStr::from_bytes(bytes))
+}
+
+#[cfg(not(unix))]
+fn os_str_from_bytes(bytes: &[u8]) -> Result<&OsStr, str::Utf8Error> {
+    str::from_utf8(bytes).map(OsStr::new)
+}
 
 /// Linker error
 #[derive(Debug, Error)]
@@ -489,7 +500,10 @@ where
                 let mut archive = Archive::new(input.as_ref());
                 while let Some(item) = archive.next_entry() {
                     let mut item = item.map_err(|e| LinkerError::IoError(path.clone(), e))?;
-                    let name = PathBuf::from(OsStr::from_bytes(item.header().identifier()));
+                    let name = PathBuf::from(
+                        os_str_from_bytes(item.header().identifier())
+                            .expect("archive member name is not valid UTF-8"),
+                    );
                     info!("linking archive item {}", name.display());
 
                     buf.clear();
@@ -645,7 +659,9 @@ fn create_target_machine(
                 // case 3.
                 info!(
                     "detected non-bpf input target {} and no explicit output --target specified, selecting `bpf'",
-                    OsStr::from_bytes(c_triple.to_bytes()).display()
+                    os_str_from_bytes(c_triple.to_bytes())
+                        .expect("LLVM target triple is not valid UTF-8")
+                        .display()
                 );
                 let c_triple = c"bpf";
                 (c_triple, llvm::target_from_triple(c_triple))
