+++ tests/btf/assembly/auxiliary/dep-exports.rs
@@ -8,7 +8,7 @@ pub fn dep_public_symbol() -> u8 {
     unsafe { core::ptr::read_volatile(core::ptr::dangling()) }
 }
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub fn dep_no_mangle() -> u8 {
     dep_public_symbol()
 }
